ra,dec float -> double

+++ src/structs.h
@@ -85,8 +85,8 @@ class Onplates : public std::vector<struct onplate> {};
 class plate {
     public:
     int tileid;
-    float tilera;
-    float tiledec;
+    double tilera;
+    double tiledec;
     double nhat[3]; // Unit vector pointing to plate
     int ipass; // Pass
     Table av_gals; // av_gals[k] : available galaxies of fiber k
